--- a/packages/story-editor/src/app/canvas/constants.js
+++ b/packages/story-editor/src/app/canvas/constants.js
@@ -13,8 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+export const RESIZE_OBSERVATION_KEY = 'resizeObservationKey';
 
-export { default as CanvasProvider } from './canvasProvider';
-export { default as useCanvas } from './useCanvas';
-export * from './constants';
-export * from './useCanvasBoundingBox';
+export const CANVAS_BOUNDING_BOX_IDS = {
+  CANVAS_CONTAINER: 'canvasContainer',
+  PAGE_CONTAINER: 'pageContainer',
+};