--- a/packages/story-editor/src/app/canvas/index.js
+++ b/packages/story-editor/src/app/canvas/index.js
@@ -16,5 +16,4 @@
 
 export { default as CanvasProvider } from './canvasProvider';
 export { default as useCanvas } from './useCanvas';
-export * from './constants';
 export * from './useCanvasBoundingBox';