--- a/packages/story-editor/src/app/index.js
+++ b/packages/story-editor/src/app/index.js
@@ -32,7 +32,12 @@ import {
   STORY_EVENTS,
 } from './story';
 
-import { useCanvas } from './canvas';
+import {
+  useCanvas,
+  useCanvasBoundingBox,
+  useCanvasBoundingBoxRef,
+  CANVAS_BOUNDING_BOX_IDS,
+} from './canvas';
 import { useLayout } from './layout';
 import { useCurrentUser } from './currentUser';
 import { useHelpCenter } from './helpCenter';
@@ -59,4 +64,7 @@ export {
   useHelpCenter,
   useUserOnboarding,
   useRightClickMenu,
+  useCanvasBoundingBox,
+  useCanvasBoundingBoxRef,
+  CANVAS_BOUNDING_BOX_IDS,
 };