… packages/story-editor/src/components/canvas/selection.js
… packages/story-editor/src/components/canvas/singleSelectionMoveable/index.js
… packages/story-editor/src/components/canvas/utils/useSnapping.js
… packages/story-editor/src/components/library/panes/media/common/test/accessibility.js
… packages/story-editor/src/components/library/test/shapes/shapePreview.js
… packages/story-editor/src/components/library/test/shapes/stickerPreview.js