@miina 2022-01-06
@littlemilkstudio Looks like the JS unit tests are still failing.

Would be great to be able to run the Karma tests here as well before merging. Currently it looks like most of the Karma tests that run, fail with this error:
https://github.com/google/web-stories-wp/runs/4718249940?check_suite_focus=true#step:6:5719

Usually (with the current Karma tests issues) the tests still run until it disconnects and generally fails but looks like in this PR something else might be going on here as well. Thoughts?